# api/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Library API",
    "description": "API documentation for the e-library project.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Development server"
    },
    {
      "url": "https://api-neocourse.dewanto.dev",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/v1/auth/signup": {
      "post": {
        "summary": "Register a new user",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "johndoe"
                  },
                  "username": {
                    "type": "string",
                    "example": "johndoe"
                  },
                  "password": {
                    "type": "string",
                    "example": "password"
                  },
                  "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "User created successfully"
                        },
                        "token": {
                          "type": "string",
                          "example": "rahasia bjirr"
                        }
                      }
                    },
                    "status": {
                      "type": "string",
                      "example": "success"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid input"
                    },
                    "status": {
                      "type": "string",
                      "example": "error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}